Fixes the unbounded read test for the nightly job. (#88)
- The current solution is the use a KeyedProcessFunction and use the
`processElement()` method to count the number of records that have been
processed.
- Whenever the expected count is reached, the method logs. Count exceed
is also logged.
- The job is terminated after a timeout of 18 minutes. 
- Since KeyedProcessFunction is used for the counting, same key is used
i.e a key named "commonKey" is added to the records to make sure they
are received together.
- To solve the problem of parallel operator, parallelism and number of
task slots have been set as 1.
---
To check for the job success:
- We check if there are any logs pertaining to exceeded count. In such a
case a failure is reported.
- In case of absence of "count-exceeded" logs, we search for the actual
count read. We then check this count with the number of rows in the BQ
table. In case of a match, the job succeeds.
- In case a less number of records are read then we raise an exception
indicating that the logs are actually not present and less number of
records being processed can be a possible cause.


/gcbrun