Adds support for `serialize()` for Primitive, `RECORD`, `UNION` and `ARRAY` datatypes. (#104)
The implementation allows actual avro records to be serialized into
storage api proto transferable protobufs.

The mapping is as follows:
1. Primitive Types - Object value is checked against the schema type
passed and converted to its respective type. e.g If the schema indicates
a `STRING` field, the value is validated to be a string and return the
passed object as `STRING`.
2.  MAP - Not supported yet.
3. UNION - Validation is done to ensure the absence of multiple not null
or only null type. In case of type of format ["null", datatype] the
datatype value is converted. `null` values are allowed in this case.
4. ARRAY - Individual values are validated and converted one by one. In
case of `null` value in array, Error is thrown.
5. RECORD - Recursive approach to convert each field items one by one. A
`MESSAGE` is returned containing the respective fields.