Add Flink Metrics for exactly once approach (#171)
Flink Metrics for Exactly Once Approach involves the following counters:
* `numberOfRecordsBufferedSinceCheckpoint` - Number of records that are
buffered by the Storage Write API using the `appendRows()` API call.
Since records are flushed at every checkpoint, this metric helps users
track where the records seen by the sink are processed to.
* `numberOfRecordsSeenByWriter` - Number of records seen by the writer
up until this time.
* `numberOfRecordsSeenByWriterSinceCheckpoint` - Number of records seen
by the writer since the past checkpoint.
* `numberOfRecordsWrittenToBigQuery` - Number of records written to
BigQuery. This metric is updated at first `write()` after a checkpoint.
Since `commit()` is an idempotent operation, there is no way to know if
the `commit()` has succeeded. In case of failure of `commit()`, the
state is reset to the previous checkpoint. The counter is updated after
checkpoint, so the metric value remains accurate.

/gcbrun

---------

Co-authored-by: Jayant Jain <jainjayant@google.com>